CL:
* Move file decompression logic to a new module.
