+++ src/engine/CMakeLists.txt
@@ -126,12 +126,7 @@ configure_tzdb()
 ####################################################################################################
 # Dependencies
 ####################################################################################################
-function(find_and_create_imported_target package_name target_name)
-    if (${package_name} STREQUAL "LibArchive" )
-        find_package(${package_name} REQUIRED)
-    else()
-        find_package(${package_name} CONFIG REQUIRED)
-    endif()
+function(action package_name target_name)
     if (${package_name}_FOUND)
         if (NOT TARGET ${target_name})
             add_library(${target_name} INTERFACE IMPORTED)
@@ -144,6 +139,16 @@ function(find_and_create_imported_target package_name target_name)
     else()
         message(FATAL_ERROR "${package_name} not found!")
     endif()
+endfunction()
+
+function(find_and_create_imported_target package_name target_name)
+    find_package(${package_name} CONFIG REQUIRED)
+    action(${package_name} ${target_name})
+endfunction()
+
+function(find_and_create_imported_target_ex package_name target_name)
+    find_package(${package_name} REQUIRED)
+    action(${package_name} ${target_name})
 endfunction()
 
 # Build test
@@ -189,7 +194,7 @@ find_and_create_imported_target("date" "date::date")
 find_and_create_imported_target("date" "date::date-tz")
 
 find_and_create_imported_target("liblzma" "liblzma::liblzma")
-find_and_create_imported_target("LibArchive" "LibArchive::LibArchive")
+find_and_create_imported_target_ex("LibArchive" "LibArchive::LibArchive")
 
 ####################################################################################################
 # Targets
@@ -226,6 +231,7 @@ add_subdirectory(${ENGINE_SOURCE_DIR}/metrics)
 add_subdirectory(${ENGINE_SOURCE_DIR}/geo)
 add_subdirectory(${ENGINE_SOURCE_DIR}/queue)
 add_subdirectory(${ENGINE_SOURCE_DIR}/apiserver)
+add_subdirectory(${ENGINE_SOURCE_DIR}/fs)
 
 #TODO isolate rxcpp
 target_link_libraries(main base cmds api CLI11::CLI11)
