+++ src/engine/source/base/CMakeLists.txt
@@ -27,9 +27,6 @@ target_link_libraries(base
     PUBLIC
     fmt::fmt-header-only
     spdlog::spdlog
-    PRIVATE
-    liblzma::liblzma
-    LibArchive::LibArchive
 )
 
 # Tests
@@ -39,13 +36,10 @@ if(ENGINE_BUILD_TEST)
 set(TEST_DIR ${CMAKE_CURRENT_LIST_DIR}/test)
 set(TEST_SRC_DIR ${CMAKE_CURRENT_LIST_DIR}/test/src)
 set(UNIT_SRC_DIR ${TEST_SRC_DIR}/unit)
-set(INPUT_FILES ${UNIT_SRC_DIR}/utils/input_files)
 
 add_library(base_test INTERFACE)
 target_include_directories(base_test INTERFACE ${TEST_DIR}/include)
 add_library(base::test ALIAS base_test)
-
-file(COPY ${INPUT_FILES} DESTINATION ${CMAKE_CURRENT_BINARY_DIR})
 
 add_executable(base_utest
     ${UNIT_SRC_DIR}/stringUtils_test.cpp
@@ -56,8 +50,6 @@ add_executable(base_utest
     ${UNIT_SRC_DIR}/utils/wazuhProtocol/wazuhRequest_test.cpp
     ${UNIT_SRC_DIR}/utils/wazuhProtocol/wazuhResponse_test.cpp
     ${UNIT_SRC_DIR}/utils/timeUtils_test.cpp
-    ${UNIT_SRC_DIR}/utils/archiveHelper_test.cpp
-    ${UNIT_SRC_DIR}/utils/xzHelper_test.cpp
     ${UNIT_SRC_DIR}/utils/threadEventDispatcher_test.cpp
     ${UNIT_SRC_DIR}/parseEvent_test.cpp
     ${UNIT_SRC_DIR}/dotPath_test.cpp
@@ -80,7 +72,6 @@ target_link_libraries(base_utest
     base
     GTest::gtest_main
     RocksDB::rocksdb
-    CURL::libcurl
 )
 gtest_discover_tests(base_utest)
