--- src/engine/source/base/test/src/unit/utils/input_files/archiveHelper/content_example1.json
@@ -1,1 +1,0 @@
-{"name":"CVE-2022-40497","offset":249309,"version":1,"inserted_at":"2023-12-14T08:34:57.850865Z","payload":{"containers":{"cna":{"affected":[{"cpes":["cpe:2.3:a:wazuh:wazuh:*:*:*:*:*:*:*:*"],"defaultStatus":"unaffected","product":"wazuh","vendor":"wazuh","versions":[{"lessThanOrEqual":"3.13.5","status":"affected","version":"3.6.1","versionType":"custom"},{"lessThanOrEqual":"4.2.7","status":"affected","version":"4.0.0","versionType":"custom"},{"lessThanOrEqual":"4.3.7","status":"affected","version":"4.3.0","versionType":"custom"}]}],"descriptions":[{"lang":"en","value":"Wazuh v3.6.1 - v3.13.5, v4.0.0 - v4.2.7, and v4.3.0 - v4.3.7 were discovered to contain an authenticated remote code execution (RCE) vulnerability via the Active Response endpoint."},{"lang":"es","value":"Se ha detectado que Wazuh versiones v3.6.1 - v3.13.5, v4.0.0 - v4.2.7, y v4.3.0 - v4.3.7, contienen una vulnerabilidad de ejecución de código remota (RCE) autenticada por medio del endpoint Active Response"}],"metrics":[{"cvssV3_1":{"attackComplexity":"LOW","attackVector":"NETWORK","availabilityImpact":"HIGH","baseScore":8.8,"baseSeverity":"HIGH","confidentialityImpact":"HIGH","integrityImpact":"HIGH","privilegesRequired":"LOW","scope":"UNCHANGED","userInteraction":"NONE","vectorString":"CVSS:3.1/AV:N/AC:L/PR:L/UI:N/S:U/C:H/I:H/A:H","version":"3.1"},"format":"CVSS"}],"problemTypes":[{"descriptions":[{"description":"NVD-CWE-noinfo","lang":"en"}]}],"providerMetadata":{"dateUpdated":"2022-09-29T15:19:41Z","orgId":"00000000-0000-4000-A000-000000000003","shortName":"nvd"},"references":[{"tags":["patch","third-party-advisory"],"url":"https://github.com/wazuh/wazuh/pull/14801"}]}},"cveMetadata":{"assignerOrgId":"8254265b-2729-46b6-b9e3-3dfca2d5bfca","assignerShortName":"mitre","cveId":"CVE-2022-40497","datePublished":"2022-09-28T00:15:09Z","dateUpdated":"2022-09-29T15:19:41Z","state":"PUBLISHED"},"dataType":"CVE_RECORD","dataVersion":"5.0"}}
