--- src/engine/source/base/test/src/unit/utils/input_files/archiveHelper/content_example2.json
@@ -1,1 +1,0 @@
-{"name":"CVE-2021-44079","offset":184230,"version":1,"inserted_at":"2023-12-13T19:15:52.679030Z","payload":{"containers":{"cna":{"affected":[{"cpes":["cpe:2.3:a:wazuh:wazuh:*:*:*:*:*:*:*:*"],"defaultStatus":"unaffected","product":"wazuh","vendor":"wazuh","versions":[{"lessThan":"4.2.5","status":"affected","version":"4.2.0","versionType":"custom"}]}],"descriptions":[{"lang":"en","value":"In the wazuh-slack active response script in Wazuh 4.2.x before 4.2.5, untrusted user agents are passed to a curl command line, potentially resulting in remote code execution."},{"lang":"es","value":"En el script de respuesta activa wazuh-slack en Wazuh versiones 4.2.x anteriores a 4.2.5, se pasan agentes de usuario no confiables a una línea de comando curl, lo que potencialmente resulta en la ejecución remota de código"}],"metrics":[{"cvssV2_0":{"accessComplexity":"LOW","accessVector":"NETWORK","authentication":"NONE","availabilityImpact":"PARTIAL","baseScore":7.5,"confidentialityImpact":"PARTIAL","integrityImpact":"PARTIAL","vectorString":"AV:N/AC:L/Au:N/C:P/I:P/A:P","version":"2.0"},"format":"CVSS"},{"cvssV3_1":{"attackComplexity":"LOW","attackVector":"NETWORK","availabilityImpact":"HIGH","baseScore":9.8,"baseSeverity":"CRITICAL","confidentialityImpact":"HIGH","integrityImpact":"HIGH","privilegesRequired":"NONE","scope":"UNCHANGED","userInteraction":"NONE","vectorString":"CVSS:3.1/AV:N/AC:L/PR:N/UI:N/S:U/C:H/I:H/A:H","version":"3.1"},"format":"CVSS"}],"problemTypes":[{"descriptions":[{"cweId":"CWE-77","description":"CWE-77","lang":"en"}]}],"providerMetadata":{"dateUpdated":"2021-12-14T20:54:21Z","orgId":"00000000-0000-4000-A000-000000000003","shortName":"nvd"},"references":[{"tags":["exploit","issue-tracking","mitigation","third-party-advisory"],"url":"https://github.com/wazuh/wazuh/issues/10858"},{"tags":["issue-tracking","patch","third-party-advisory"],"url":"https://github.com/wazuh/wazuh/issues/10858#issuecomment-991118254"},{"tags":["patch","third-party-advisory"],"url":"https://github.com/wazuh/wazuh/pull/10809"}]}},"cveMetadata":{"assignerOrgId":"8254265b-2729-46b6-b9e3-3dfca2d5bfca","assignerShortName":"mitre","cveId":"CVE-2021-44079","datePublished":"2021-11-22T07:15:07Z","dateUpdated":"2021-12-14T20:54:21Z","state":"PUBLISHED"},"dataType":"CVE_RECORD","dataVersion":"5.0"}}
